shuffling works
